move queries helpers in helpers directory
