+++ lib/queries/batchInsert.js
@@ -1,6 +1,6 @@
-import valueSubQuery from './valueSubQuery';
-import batchParameter from './batchParameter';
-import returningQuery from './returningQuery';
+import valueSubQuery from './helpers/valueSubQuery';
+import batchParameter from './helpers/batchParameter';
+import returningQuery from './helpers/returningQuery';
 
 
 export default ({ table, writableFields, returnFields }) => {
