+++ lib/queries/batchRemove.js
@@ -1,8 +1,8 @@
-import batchParameter from './batchParameter';
-import sanitizeIdentifier from './sanitizeIdentifier';
+import batchParameter from './helpers/batchParameter';
+import sanitizeIdentifier from './helpers/sanitizeIdentifier';
 import combineLiterals from '../utils/combineLiterals';
-import whereQuery from './whereQuery';
-import returningQuery from './returningQuery';
+import whereQuery from './helpers/whereQuery';
+import returningQuery from './helpers/returningQuery';
 
 export default ({ table, returnFields, primaryKey = ['id'] }) => {
     const returning = returningQuery(returnFields);
