+++ lib/queries/batchUpsert.js
@@ -1,5 +1,5 @@
-import valueSubQuery from './valueSubQuery';
-import batchParameter from './batchParameter';
+import valueSubQuery from './helpers/valueSubQuery';
+import batchParameter from './helpers/batchParameter';
 
 export default ({
     table,
