+++ lib/queries/crud.js
@@ -1,11 +1,11 @@
-import removeOneQuery from '../queries/removeOne';
-import insertOneQuery from '../queries/insertOne';
-import selectOneQuery from '../queries/selectOne';
-import selectQuery from '../queries/select';
-import updateOneQuery from '../queries/updateOne';
-import batchInsertQuery from '../queries/batchInsert';
-import batchRemoveQuery from '../queries/batchRemove';
-import countAllQuery from '../queries/countAll';
+import removeOneQuery from './removeOne';
+import insertOneQuery from './insertOne';
+import selectOneQuery from './selectOne';
+import selectQuery from './select';
+import updateOneQuery from './updateOne';
+import batchInsertQuery from './batchInsert';
+import batchRemoveQuery from './batchRemove';
+import countAllQuery from './countAll';
 
 export default function ({
     table,
