+++ lib/queries/helpers/batchParameter.js
@@ -1,6 +1,6 @@
 import sanitizeParameter from './sanitizeParameter';
-import arrayToLitteral from '../utils/arrayToLitteral';
-import addSuffix from '../utils/addSuffix';
+import arrayToLitteral from '../../utils/arrayToLitteral';
+import addSuffix from '../../utils/addSuffix';
 
 export default function (fields) {
     const sanitize = sanitizeParameter(arrayToLitteral(fields, null));
