+++ lib/queries/helpers/flattenParameters.js
