+++ lib/queries/helpers/sanitizeIdentifier.js
@@ -1,8 +1,8 @@
-import pipe from '../utils/pipe';
-import curry from '../utils/curry';
+import pipe from '../../utils/pipe';
+import curry from '../../utils/curry';
+import checkLiteralKeys from '../../utils/checkLiteralKeys';
+import ensureIsSet from '../../utils/ensureIsSet';
 import sanitizeParameter from './sanitizeParameter';
-import checkLiteralKeys from '../utils/checkLiteralKeys';
-import ensureIsSet from '../utils/ensureIsSet';
 
 export const normalize = identifierFields => id =>
     (typeof id === 'object' ? id : { [identifierFields[0]]: id });
