+++ lib/queries/helpers/sanitizeIdentifier.spec.js
