+++ lib/queries/helpers/sanitizeParameter.js
@@ -1,5 +1,5 @@
-import curry from '../utils/curry';
-import arrayToLitteral from '../utils/arrayToLitteral';
+import curry from '../../utils/curry';
+import arrayToLitteral from '../../utils/arrayToLitteral';
 import { getFieldType } from './whereQuery';
 
 function getTrueFieldName(fieldName, fields) {
