+++ lib/queries/helpers/valueSubQuery.js
@@ -1,4 +1,4 @@
-import curry from '../utils/curry';
+import curry from '../../utils/curry';
 
 function valueSubQuery(writableFields, suffix) {
     return writableFields
