+++ lib/queries/helpers/whereQuery.js
@@ -1,4 +1,4 @@
-import middleware from '../utils/middleware';
+import middleware from '../../utils/middleware';
 
 export function getFieldPlaceHolder(field, value) {
     const type = Array.isArray(value) ? 'IN' : value;
