+++ lib/queries/insertOne.js
@@ -1,6 +1,6 @@
-import valueSubQuery from './valueSubQuery';
-import sanitizeParameter from './sanitizeParameter';
-import returningQuery from './returningQuery';
+import valueSubQuery from './helpers/valueSubQuery';
+import sanitizeParameter from './helpers/sanitizeParameter';
+import returningQuery from './helpers/returningQuery';
 
 export default ({ table, writableFields, returnFields }) => {
     const returning = returningQuery(returnFields);
