+++ lib/queries/remove.js
@@ -1,7 +1,7 @@
-import sanitizeIdentifier from './sanitizeIdentifier';
-import sanitizeParameter from './sanitizeParameter';
-import whereQuery from './whereQuery';
-import returningQuery from './returningQuery';
+import sanitizeIdentifier from './helpers/sanitizeIdentifier';
+import sanitizeParameter from './helpers/sanitizeParameter';
+import whereQuery from './helpers/whereQuery';
+import returningQuery from './helpers/returningQuery';
 
 export default ({ table, filterFields, returnFields }, _one = false) => {
     const returning = returningQuery(returnFields);
