+++ lib/queries/select.js
@@ -1,6 +1,6 @@
-import whereQuery from './whereQuery';
-import sanitizeParameter from './sanitizeParameter';
-import flattenParameters from './flattenParameters';
+import whereQuery from './helpers/whereQuery';
+import sanitizeParameter from './helpers/sanitizeParameter';
+import flattenParameters from './helpers/flattenParameters';
 
 export default ({
     table,
