+++ lib/queries/selectByOrderedFieldValues.js
@@ -1,5 +1,5 @@
-import sanitizeParameter from './sanitizeParameter';
-import flattenParameters from './flattenParameters';
+import sanitizeParameter from './helpers/sanitizeParameter';
+import flattenParameters from './helpers/flattenParameters';
 
 export default ({ table, selectorField, returnFields = ['*'] }) => (values) => {
     if (!Array.isArray(values)) {
