+++ lib/queries/selectOne.js
@@ -1,5 +1,5 @@
-import whereQuery from './whereQuery';
-import sanitizeIdentifier from './sanitizeIdentifier';
+import whereQuery from './helpers/whereQuery';
+import sanitizeIdentifier from './helpers/sanitizeIdentifier';
 
 export default ({ table, primaryKey = 'id', returnFields = ['*'] }) => {
     const select = returnFields.join(', ');
