+++ lib/queries/selectOne.spec.js
@@ -1,6 +1,6 @@
 import { assert } from 'chai';
 
-import selectOneQuerier from '../../lib/queries/selectOne';
+import selectOneQuerier from './selectOne';
 
 describe('QUERY selectOne', () => {
     it('should generate sql and parameter for selecting one entity', () => {
