+++ lib/queries/update.js
@@ -1,7 +1,7 @@
-import whereQuery from './whereQuery';
-import sanitizeParameter from './sanitizeParameter';
-import sanitizeIdentifier from './sanitizeIdentifier';
-import returningQuery from './returningQuery';
+import whereQuery from './helpers/whereQuery';
+import sanitizeParameter from './helpers/sanitizeParameter';
+import sanitizeIdentifier from './helpers/sanitizeIdentifier';
+import returningQuery from './helpers/returningQuery';
 import addSuffix from '../utils/addSuffix';
 
 export default ({
