+++ lib/queries/upsertOne.js
@@ -1,7 +1,7 @@
-import valueSubQuery from './valueSubQuery';
-import sanitizeParameter from './sanitizeParameter';
-import sanitizeIdentifier from './sanitizeIdentifier';
-import returningQuery from './returningQuery';
+import valueSubQuery from './helpers/valueSubQuery';
+import sanitizeParameter from './helpers/sanitizeParameter';
+import sanitizeIdentifier from './helpers/sanitizeIdentifier';
+import returningQuery from './helpers/returningQuery';
 
 export default ({
     table,
